+++ CHANGELOG.md
@@ -4,4 +4,4 @@ List of changes for every update.
 
 ## TOC
   ### Top-level Github
-- React native demo app to show how to use the `fbt-rn-android-native` module
+- React native demo app to show how to use the `react-native-fbt` module